rename config to properties clean up
